feat: Add payment system

+++ WrocRide/Entities/Driver.cs
@@ -6,7 +6,7 @@ namespace WrocRide.Entities
     {
         public int Id { get; set; }
         public float? Rating { get; set; }
-        public float Pricing { get; set; }
+        public decimal Pricing { get; set; }
         public DriverStatus DriverStatus { get; set; }
         public int UserId { get; set; }
         public virtual User User { get; set; }
